Rename the project from `withingsslack` to `slackhealthbot`.
it's not limited only to withings: fitbit is also supported.

Reorder imports using isort.
