+++ README.md
@@ -60,7 +60,7 @@ curl --location 'http://your-server/withings-notification-webhook/' \
 
 You can find your userid in the database file:
 ```
-sqlite3 -header -column /path/to/withingsslack.db \
+sqlite3 -header -column /path/to/slackhealthbot.db \
   "select
      slack_alias,
      withings_users.oauth_userid as withings_userid,
