+++ alembic/env.py
@@ -1,12 +1,10 @@
 from logging.config import fileConfig
 
-from sqlalchemy import engine_from_config
-from sqlalchemy import pool
+from sqlalchemy import engine_from_config, pool
 
 from alembic import context
-
-from withingsslack.database import models as db_models
-from withingsslack.database.connection import connection_url
+from slackhealthbot.database import models as db_models
+from slackhealthbot.database.connection import connection_url
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
