+++ alembic/versions/2023_05_13_1218-e1da2ea0086f_initial_revision.py
@@ -5,9 +5,9 @@ Revises:
 Create Date: 2023-05-13 12:18:07.552386
 
 """
-from alembic import op
 import sqlalchemy as sa
 
+from alembic import op
 
 # revision identifiers, used by Alembic.
 revision = "e1da2ea0086f"
