+++ alembic/versions/2023_05_13_1254-d1a96aed325a_use_declarative_mapping.py
@@ -5,9 +5,9 @@ Revises: e1da2ea0086f
 Create Date: 2023-05-13 12:54:04.773019
 
 """
-from alembic import op
 import sqlalchemy as sa
 
+from alembic import op
 
 # revision identifiers, used by Alembic.
 revision = "d1a96aed325a"
