+++ requirements/dev.txt
@@ -1,5 +1,6 @@
 -r prod.txt
 ruff==0.0.263
 black==23.3.0
+isort==5.12.0
 pytest==7.3.1
 pytest-cov==4.0.0
