+++ scripts/codecheck.sh
@@ -1,5 +1,6 @@
-for project in withingsslack alembic tests
+for project in slackhealthbot alembic tests
 do
   black $project
   ruff check $project
+  isort $project
 done
