+++ scripts/run_tests.sh
@@ -1,4 +1,4 @@
 rm -rf reports
-python -m pytest --cov=withingsslack --cov-report=xml --cov-report=html --junitxml="reports/junit.xml" tests
+python -m pytest --cov=slackhealthbot --cov-report=xml --cov-report=html --junitxml="reports/junit.xml" tests
 mkdir -p reports
 mv coverage.xml htmlcov reports/.
