+++ slackhealthbot/services/exceptions.py
